# docs/swagger.yaml
definitions:
  modelsViewApi.BadRequestError:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  modelsViewApi.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  modelsViewApi.ProtestComplete:
    properties:
      comment:
        type: string
      resPoints:
        type: integer
    required:
    - resPoints
    type: object
  modelsViewApi.ProtestCreate:
    properties:
      comment:
        type: string
      judge-id:
        type: string
      protestee:
        type: integer
      protestor:
        type: integer
      reviewDate:
        type: string
      ruleNum:
        type: integer
      witnesses:
        items:
          type: integer
        type: array
    required:
    - protestee
    - protestor
    - reviewDate
    - ruleNum
    - witnesses
    type: object
  modelsViewApi.ProtestCrewFormData:
    properties:
      class:
        type: string
      id:
        type: string
      name:
        type: integer
      rating-id:
        type: string
      role:
        type: string
    type: object
  modelsViewApi.ProtestFormData:
    properties:
      comment:
        type: string
      id:
        type: string
      judge-id:
        type: string
      race-id:
        type: string
      rating-id:
        type: string
      review-date:
        type: string
      rule-num:
        type: integer
      status:
        type: string
    type: object
  modelsViewApi.ProtestInput:
    properties:
      comment:
        type: string
      judge-id:
        type: string
      reviewDate:
        type: string
      ruleNum:
        type: integer
      status:
        type: integer
    required:
    - reviewDate
    - ruleNum
    type: object
  modelsViewApi.ProtestParticipantAttachInput:
    properties:
      role:
        type: integer
      sailNum:
        type: integer
    required:
    - role
    - sailNum
    type: object
...
info:
  contact: {}
paths:
...
  /api/ratings/{ratingID}/races/{raceID}/protests:
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      responses:
        "200":
          description: Список протестов успешно получен
          schema:
            items:
              $ref: '#/definitions/modelsViewApi.ProtestFormData'
            type: array
        "404":
          description: Рейтинг или гонка не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить все протесты для указанного рейтинга и гонки
      tags:
      - Protests
    post:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Данные для создания протеста
        in: body
        name: protestInput
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.ProtestCreate'
      responses:
        "201":
          description: Протест успешно создан
          schema:
            $ref: '#/definitions/modelsViewApi.ProtestFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг или гонка не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Создать новый протест
      tags:
      - Protests
  /api/ratings/{ratingID}/races/{raceID}/protests/{protestID}:
    delete:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Уникальный идентификатор протеста
        format: uuid
        in: path
        name: protestID
        required: true
        type: string
      responses:
        "204":
          description: Протест успешно удален
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Протест, рейтинг или гонка не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Удалить протест
      tags:
      - Protests
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Уникальный идентификатор протеста
        format: uuid
        in: path
        name: protestID
        required: true
        type: string
      responses:
        "200":
          description: Информация о протесте
          schema:
            $ref: '#/definitions/modelsViewApi.ProtestFormData'
        "404":
          description: Протест, рейтинг или гонка не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить информацию о протесте
      tags:
      - Protests
    put:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Уникальный идентификатор протеста
        format: uuid
        in: path
        name: protestID
        required: true
        type: string
      - description: Данные для обновления протеста
        in: body
        name: protestInput
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.ProtestInput'
      responses:
        "200":
          description: Протест успешно обновлен
          schema:
            $ref: '#/definitions/modelsViewApi.ProtestFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Протест, рейтинг или гонка не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Обновить информацию о протесте
      tags:
      - Protests
  /api/ratings/{ratingID}/races/{raceID}/protests/{protestID}/complete:
    patch:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Уникальный идентификатор протеста
        format: uuid
        in: path
        name: protestID
        required: true
        type: string
      - description: Данные для завершения рассмотрения протеста
        in: body
        name: protestComplete
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.ProtestComplete'
      responses:
        "200":
          description: Рассмотрение протеста успешно завершено
          schema:
            $ref: '#/definitions/modelsViewApi.ProtestFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Протест, рейтинг или гонка не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Завершить рассмотрение протеста
      tags:
      - Protests
  /api/ratings/{ratingID}/races/{raceID}/protests/{protestID}/members:
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Уникальный идентификатор протеста
        format: uuid
        in: path
        name: protestID
        required: true
        type: string
      responses:
        "200":
          description: Список команд-участников протеста
          schema:
            items:
              $ref: '#/definitions/modelsViewApi.ProtestCrewFormData'
            type: array
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Протест, гонка или команды не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить информацию о всех командах-участниках протеста
      tags:
      - Protest Members
    post:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Уникальный идентификатор протеста
        format: uuid
        in: path
        name: protestID
        required: true
        type: string
      - description: Данные для добавления команды-участника
        in: body
        name: protestParticipantAttachInput
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.ProtestParticipantAttachInput'
      responses:
        "201":
          description: Команда-участник успешно добавлена
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Протест, гонка или команда не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Добавить команду-участника протеста
      tags:
      - Protest Members
  /api/ratings/{ratingID}/races/{raceID}/protests/{protestID}/members/{crewSailNum}:
    delete:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Уникальный идентификатор протеста
        format: uuid
        in: path
        name: protestID
        required: true
        type: string
      - description: Номер паруса команды
        format: int
        in: path
        name: crewSailNum
        required: true
        type: integer
      responses:
        "204":
          description: Команда-участник успешно удалена из протеста
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Протест, гонка или команда не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Удалить команду-участника из протеста
      tags:
      - Protest Members
...
swagger: "2.0"